requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
openpyxl>=3.1.0
xlrd>=2.0.0
numpy>=1.26.0
scipy>=1.12.0
jinja2>=3.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
# Authentication
PyJWT>=2.8.0
Werkzeug>=3.0.0
# AI extraction
anthropic>=0.30.0
# Document parsing
pypdf>=3.0.0
pdfplumber>=0.10.0
python-pptx>=0.6.21
python-docx>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
# PDF export
weasyprint>=60.0
